progress on adopting the benchmarks!

+++ run_benchmarks.py
@@ -8,6 +8,9 @@ from pathlib import Path
 import sys
 from multiprocessing import Process
 import time
+import json
+
+BENCHMARKS = "benchmarks"
 
 TIMEOUT = 2  # seconds
 
@@ -16,20 +19,28 @@ NO_GEN_PARTITIONS = 100
 PROPAGATE = False
 
 DEBUG = False
-TRACE = False
+TRACE = True
 # [Cs, Cb]  [Cs, CbGen, CmGen]
 HEURISTICS = [Cs, CbGen]
 # [no_generalization, linear_generalize_state_conflict, linear_generalize_state_binary, exponential_generalize_state]
 GENERALIZATIONS = [None, linear_generalization, binary_generalization, exponential_generalization]
 
-MIN_ITERATION = int(sys.argv[1])
+if len(sys.argv) > 1:
+    MIN_ITERATION = int(sys.argv[1])
+else:
+    MIN_ITERATION = 0
 iteration = 0
-BENCHMARK_FOLDER = Path("benchmarks")
+BENCHMARK_FOLDER = Path(BENCHMARKS)
+
+with open(BENCHMARKS + "/lambdas.json", "r") as f:
+    LAMBDAS = json.load(f)
 
 def run_benchmark(M: Model, do_propagate: bool, heuristics, used_heuristic, generalizations, used_generalization, 
                    print_ : bool = True, assert_: bool = True, loop_check: bool = True):
     try:
         result = adjointPDRdown(M,do_propagate,heuristics,used_heuristic, generalizations, used_generalization, print_, assert_, loop_check)
+    except MemoryError:
+        print("M/O", end=" ")
     except:
         print("CRASHED", end=" ")
     print(result[0], end=" ")
@@ -40,32 +51,35 @@ def hname(gen: Callable | None) -> str:
     return "None" if gen is None else gen.__name__
 
 for file in sorted(BENCHMARK_FOLDER.iterdir()):
-    if file.is_file():
+    if file.is_file() and str(file.suffix) == ".pm":
+        if TRACE:
+            print("Opening:", str(file))
         with file.open("r") as f:
-            max_prob = Fraction(f.readline().replace("//", "").replace(" ", ""))
-            model = Model.from_prism_file(ctx, str(file), max_prob, False, NO_GEN_PARTITIONS)
-            
-            for heur in HEURISTICS:
-                for gen in GENERALIZATIONS:
-                    if iteration < MIN_ITERATION:
-                        continue
+            for lambda_float in LAMBDAS[str(file.name)]:
+                lambda_ = Fraction(lambda_float)
+                model = Model.from_prism_file(ctx, str(file), lambda_, False, NO_GEN_PARTITIONS, bad_label="goal")
+                
+                for heur in HEURISTICS:
+                    for gen in GENERALIZATIONS:
+                        if iteration < MIN_ITERATION:
+                            continue
 
-                    if TRACE:
-                        print("**NOW RUNNING**", str(file), float(max_prob), hname(heur), hname(gen), "...")
+                        if TRACE:
+                            print("**NOW RUNNING**", str(file), lambda_float, hname(heur), hname(gen), "...")
+                        
+                        p = Process(target=run_benchmark, args=(model, False,[heur], heur, [gen], gen, DEBUG, DEBUG, DEBUG))
+                        start = time.perf_counter()
+                        p.start()
+                        p.join(TIMEOUT)
+                        if p.is_alive():
+                            p.terminate()
+                            p.join()
+                            print(iteration, str(file), lambda_float,hname(heur), hname(gen), "T/O")
+                        else:
+                            elapsed = time.perf_counter() - start
+                            print(iteration, str(file), lambda_float, hname(heur), hname(gen), f"{elapsed:.3f}")
+                        iteration += 1
                     
-                    p = Process(target=run_benchmark, args=(model, False,[heur], heur, [gen], gen, DEBUG, DEBUG, DEBUG))
-                    start = time.perf_counter()
-                    p.start()
-                    p.join(TIMEOUT)
-                    if p.is_alive():
-                        p.terminate()
-                        p.join()
-                        print(iteration, str(file), float(max_prob),hname(heur), hname(gen), "T/O")
-                    else:
-                        elapsed = time.perf_counter() - start
-                        print(iteration, str(file), float(max_prob), hname(heur), hname(gen), f"{elapsed:.3f}")
-                    iteration += 1
-                    
-
+print("Done.")
